enable setup draft GH multistep and clean up.
