+++ src/commands/runDraftTool/draftCreateMultiStep/MultiStepInput.ts
@@ -3,10 +3,11 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { QuickPickItem, window, Disposable, CancellationToken, QuickInputButton, QuickInput, ExtensionContext, QuickInputButtons, Uri } from 'vscode';
+import { QuickPickItem, window, CancellationToken, ExtensionContext } from 'vscode';
 import { buildCreateCommand, buildCreateConfig } from '../helper/draftCommandBuilder';
 import { runDraftCommand } from '../helper/runDraftHelper';
 import { reporter } from '../../../utils/reporter';
+import { MultiStepInput } from '../../../utils/multiStepWizardWrapper';
 
 /**
  * A multi-step input using window.createQuickPick() and window.createInputBox().
@@ -84,7 +85,6 @@ export async function multiStepInput(context: ExtensionContext, destination: str
 			shouldResume: shouldResume
 		});
         return (input: MultiStepInput) => inputApplicationName(input, state);
-		// return (input: MultiStepInput) => inputName(input, state);
 	}
 
 	async function inputApplicationName(input: MultiStepInput, state: Partial<State>) {
@@ -141,13 +141,6 @@ export async function multiStepInput(context: ExtensionContext, destination: str
 		return name === 'vscode' ? 'Name not unique' : undefined;
 	}
 
-	async function getAvailableRuntimes(resourceGroup: QuickPickItem | string, token?: CancellationToken): Promise<QuickPickItem[]> {
-		// ...retrieve...
-		await new Promise(resolve => setTimeout(resolve, 1000));
-		return ['Node 8.9', 'Node 6.11', 'Node 4.5']
-			.map(label => ({ label }));
-	}
-
 	const state = await collectInputs();
 
     const language = state.language;
@@ -167,182 +160,4 @@ export async function multiStepInput(context: ExtensionContext, destination: str
     }
 
 	window.showInformationMessage(`Draft Message - '${result}'`);
-}
-
-
-// -------------------------------------------------------
-// Helper code that wraps the API for the multi-step case.
-// -------------------------------------------------------
-
-
-class InputFlowAction {
-	static back = new InputFlowAction();
-	static cancel = new InputFlowAction();
-	static resume = new InputFlowAction();
-}
-
-type InputStep = (input: MultiStepInput) => Thenable<InputStep | void>;
-
-interface QuickPickParameters<T extends QuickPickItem> {
-	title: string;
-	step: number;
-	totalSteps: number;
-	items: T[];
-	activeItem?: T;
-	placeholder: string;
-	buttons?: QuickInputButton[];
-	shouldResume: () => Thenable<boolean>;
-}
-
-interface InputBoxParameters {
-	title: string;
-	step: number;
-	totalSteps: number;
-	value: string;
-	prompt: string;
-	validate: (value: string) => Promise<string | undefined>;
-	buttons?: QuickInputButton[];
-	shouldResume: () => Thenable<boolean>;
-}
-
-class MultiStepInput {
-
-	static async run<T>(start: InputStep) {
-		const input = new MultiStepInput();
-		return input.stepThrough(start);
-	}
-
-	private current?: QuickInput;
-	private steps: InputStep[] = [];
-
-	private async stepThrough<T>(start: InputStep) {
-		let step: InputStep | void = start;
-		while (step) {
-			this.steps.push(step);
-			if (this.current) {
-				this.current.enabled = false;
-				this.current.busy = true;
-			}
-			try {
-				step = await step(this);
-			} catch (err) {
-				if (err === InputFlowAction.back) {
-					this.steps.pop();
-					step = this.steps.pop();
-				} else if (err === InputFlowAction.resume) {
-					step = this.steps.pop();
-				} else if (err === InputFlowAction.cancel) {
-					step = undefined;
-				} else {
-					throw err;
-				}
-			}
-		}
-		if (this.current) {
-			this.current.dispose();
-		}
-	}
-
-	async showQuickPick<T extends QuickPickItem, P extends QuickPickParameters<T>>({ title, step, totalSteps, items, activeItem, placeholder, buttons, shouldResume }: P) {
-		const disposables: Disposable[] = [];
-		try {
-			return await new Promise<T | (P extends { buttons: (infer I)[] } ? I : never)>((resolve, reject) => {
-				const input = window.createQuickPick<T>();
-				input.title = title;
-				input.step = step;
-				input.totalSteps = totalSteps;
-				input.placeholder = placeholder;
-				input.items = items;
-				if (activeItem) {
-					input.activeItems = [activeItem];
-				}
-				input.buttons = [
-					...(this.steps.length > 1 ? [QuickInputButtons.Back] : []),
-					...(buttons || [])
-				];
-				disposables.push(
-					input.onDidTriggerButton(item => {
-						if (item === QuickInputButtons.Back) {
-							reject(InputFlowAction.back);
-						} else {
-							resolve(<any>item);
-						}
-					}),
-					input.onDidChangeSelection(items => resolve(items[0])),
-					input.onDidHide(() => {
-						(async () => {
-							reject(shouldResume && await shouldResume() ? InputFlowAction.resume : InputFlowAction.cancel);
-						})()
-							.catch(reject);
-					})
-				);
-				if (this.current) {
-					this.current.dispose();
-				}
-				this.current = input;
-				this.current.show();
-			});
-		} finally {
-			disposables.forEach(d => d.dispose());
-		}
-	}
-
-	async showInputBox<P extends InputBoxParameters>({ title, step, totalSteps, value, prompt, validate, buttons, shouldResume }: P) {
-		const disposables: Disposable[] = [];
-		try {
-			return await new Promise<string | (P extends { buttons: (infer I)[] } ? I : never)>((resolve, reject) => {
-				const input = window.createInputBox();
-				input.title = title;
-				input.step = step;
-				input.totalSteps = totalSteps;
-				input.value = value || '';
-				input.prompt = prompt;
-				input.buttons = [
-					...(this.steps.length > 1 ? [QuickInputButtons.Back] : []),
-					...(buttons || [])
-				];
-				let validating = validate('');
-				disposables.push(
-					input.onDidTriggerButton(item => {
-						if (item === QuickInputButtons.Back) {
-							reject(InputFlowAction.back);
-						} else {
-							resolve(<any>item);
-						}
-					}),
-					input.onDidAccept(async () => {
-						const value = input.value;
-						input.enabled = false;
-						input.busy = true;
-						if (!(await validate(value))) {
-							resolve(value);
-						}
-						input.enabled = true;
-						input.busy = false;
-					}),
-					input.onDidChangeValue(async text => {
-						const current = validate(text);
-						validating = current;
-						const validationMessage = await current;
-						if (current === validating) {
-							input.validationMessage = validationMessage;
-						}
-					}),
-					input.onDidHide(() => {
-						(async () => {
-							reject(shouldResume && await shouldResume() ? InputFlowAction.resume : InputFlowAction.cancel);
-						})()
-							.catch(reject);
-					})
-				);
-				if (this.current) {
-					this.current.dispose();
-				}
-				this.current = input;
-				this.current.show();
-			});
-		} finally {
-			disposables.forEach(d => d.dispose());
-		}
-	}
 }
